Rename CloudInviteUsersHint component file
